:bug: fix form_single_domain_2D to stop bursting loop
Check if domain is a single circular and then calculate modest_radius.
The difficulty is the difference between modest_radius and min_shell_
radius. The modest radius depends on the diffusion coefficient of the
particle in the domain, but the min_shell_radius does not. This makes a
gap between min_shell_radius and modest_radius and may cause an infinite
loop.